protobuf 强转
