+++ _example/helloworld/rpccli/main.go
@@ -6,18 +6,18 @@ import (
 
 	"github.com/fengjx/luchen/env"
 
-	"github.com/fengjx/luchen/example/greet/pb"
+	"github.com/fengjx/luchen/example/helloworld/pb"
 )
 
 func init() {
-	if env.IsLocal() {
+	if env.IsDev() {
 		// 可以设置环境变量 LUCHEN_ETCD_ADDRESS 指定 etcd 地址
-		env.SetDefaultEtcdAddress([]string{"host.etcd.dev:2379"})
+		env.SetDefaultEtcdAddress([]string{"192.168.6.129:2379"})
 	}
 }
 
 func main() {
-	greeterClient := pb.NewGreeterService("featgrpc")
+	greeterClient := pb.NewGreeterService("greet")
 	helloResp, err := greeterClient.SayHello(context.Background(), &pb.HelloReq{
 		Name: "fengjx",
 	})
